Update 4
•	Keep track of connected users in the chat – this should be working
o	A user should have a unique identiﬁer (e.g. email or a username), otherwise they should not be allowed to join – this should work
•	A user also needs to be able to leave – we have a function for it, but we do not know if it actually works
•	Keep a record of all the messages – this we are missing!!!
•	Let users send a message to the chat – we have the function for it, but not sure if it works properly
o	A message need to contain: A message, a send date and an author – we have it in the function, but I am not sure if it is being displayed
•	Search within messages – we have the ‘keyword’ as part of function (indexOf), but we are not sure it works.
o	By content and by author – included in code, but not sure it works.
•	Censor messages containing certain words – this function we have, however it can only detect the bad word when it is not in a sentence. Also, it does not work when we add the message function.
o	Censored should still be saved – do not know how to do this!!!